Fix to be compatible with CommunityPatch mod (#1)
As of 0.1.0-rc2 CommunityPatch modifies the FirstPhase storyline quests to show that they will fail after 10 game years. It does this by using the QuestDueTime property on top of how vanilla failed them which was through an unrelated tick handler. The result is the same, but the gui will display the days remaining.

We therefore modify them in the opposite way to remove the QuestDueTime setting. As CommunityPatch does this in an HourlyTick handler after the quest starts, we do our changes in the 2nd HourlyTick after quest start to reverse it.