crash the game when we fail to modify quests
It's probably more important for people to know when the mod has failed to do it's work rather than to avoid the annoyance of a crash. At least if it crashes the person knows something needs to be changed, but if it silently fails they will continue the game, the story quests may timeout, and they may not have a save to go back to.